Construct dict directly
No need to construct and then modify when you can just use a
dict-literal. Also, not good practice to call things "dict" when that's
the built-in class name (I would rather name the method as_dict).

Also call str rather than using the __str__ method (as a rough rule of
thumb, the dunder methods should not be called directly).