Tidy up usage of argparse
Replace all uses of optparse with argparse, and ensure parse_args is
always called with args passed into main (for ease of testing).

Tidy up use of argparse: no need to specify action=store (it's the
default), or the destination var (it's taken from the long-form of the
argument, converting "-" to "_" and excluding the leading "--" if any).
Report user errors with parser.error (which handles showing the usage
info on stderr, and exiting with an appropriate exit-code of 2).