Handle config loading errors better
As this is external untrusted input we should explicitly convert the
incoming type (JSON objects can store values other than strings). This
also implies we should catch ValueError for invalid conversions.

There's also no guarantee that a stored configuration contains all the
correct keys, so we should catch KeyError for this case. Finally, the
JSONDecodeError is a subclass of ValueError so it's redundant in the
exception list.